added service level logic to each API call

+++ app/api/v1/event.py
@@ -3,12 +3,19 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 from app import crud, schemas, database, models
 from app.database import get_db
+from app.api.v1.dependencies import get_current_user
+from app.crud import get_user_by_email
 
 router = APIRouter()
 
 # Create a new event
 @router.post("/", response_model=schemas.EventOut,)
-def create_event(event: schemas.EventCreate, db: Session = Depends(get_db)):
+def create_event(event: schemas.EventCreate, db: Session = Depends(get_db), current_user=Depends(get_current_user)):
+    user = get_user_by_email(db, current_user["email"])
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    if user.role not in ["admin", "service"]:
+        raise HTTPException(status_code=403, detail="Forbidden")
     return crud.create_event(db, event)
 
 # Get an event by ID
@@ -26,7 +33,12 @@ def get_all_events(db: Session = Depends(get_db)):
 
 # Delete an event
 @router.delete("/event/{event_id}")
-def delete_event(event_id: int, db: Session = Depends(database.get_db)):
+def delete_event(event_id: int, db: Session = Depends(database.get_db), current_user=Depends(get_current_user)):
+    user = get_user_by_email(db, current_user["email"])
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    if user.role not in ["admin", "service"]:
+        raise HTTPException(status_code=403, detail="Forbidden")
     result = crud.delete_event(db, event_id)
 
     if result == "not_found":
